requirements.txt
SQLAlchemy~=2.0.38
fastapi~=0.115.8
pydantic-settings~=2.8.0
pydantic~=2.10.6
python-dotenv~=1.0.1
alembic~=1.14.1
docker-py~=1.10.6
pypiwin32
docker~=7.1.0
PyYAML~=6.0.2
kubernetes~=32.0.1